test(services): Move test to CoreEnvironment
Since this is testing the logic in CoreEnvironment::build and down.
